refactor: split out nonce-management into multiple files
