+++ packages/core/src/internal/execution/nonce-management/get-max-nonce-used-by-sender.ts
@@ -1,0 +1,28 @@
+import { getPendingNonceAndSender } from "../../views/execution-state/get-pending-nonce-and-sender";
+import { DeploymentState } from "../types/deployment-state";
+
+export function getMaxNonceUsedBySender(deploymentState: DeploymentState): {
+  [sender: string]: number;
+} {
+  const nonces: {
+    [sender: string]: number;
+  } = {};
+
+  for (const executionState of Object.values(deploymentState.executionStates)) {
+    const pendingNonceAndSender = getPendingNonceAndSender(executionState);
+
+    if (pendingNonceAndSender === undefined) {
+      continue;
+    }
+
+    const { sender, nonce } = pendingNonceAndSender;
+
+    if (nonces[sender] === undefined) {
+      nonces[sender] = nonce;
+    } else {
+      nonces[sender] = Math.max(nonces[sender], nonce);
+    }
+  }
+
+  return nonces;
+}
